requirements.txt
fastapi
uvicorn
crewai[tools]
langchain-groq
langsmith
sqlmodel
python-dotenv
requests